==> Microbiome_Analysis/5_stats/00_wrapper.Rmd <==
---
title: "Microplastic - Microbiome Analysis"
author: "Jan Waelchli"
date: '`r Sys.Date()`'
geometry: margin=2cm
output:
  pdf_document:
    toc: yes
    toc_depth: 2
---

```{r setup, include=FALSE}

#This file is used to knitr all chapter together as one pdf!

knitr::opts_chunk$set(echo = TRUE, fig.align="center")
options(knitr.duplicate.label = "allow", tinytex.verbose = TRUE)

## set source
library(rstudioapi)
setwd(dirname(getActiveDocumentContext()$path))

```

```{r, child=c('01_Import_Normalization.Rmd', '02_Taxa_Analysis.Rmd', '03_Alpha_Diversity.Rmd', '04_Beta_Diversity.Rmd', '05_DAA.Rmd')}
#, '06_Correlations.Rmd'

#By knitting this script you receive one pdf out of all child Rmd-documents.

```
